update router to take a long instead as id like backend

+++ android/app/src/main/java/ch/morgias/cookgenda/android/Screen.kt
@@ -11,7 +11,7 @@ sealed class Screen(val route: String) {
     object PlanningListScreen : Screen(route = "planning")
     object ShoppingListResumesScreen : Screen(route = "shoppingList")
     object ShoppingListScreen: Screen(route = "shoppingList/{shoppingListId}"){
-        fun withShoppingListId(shoppingListId: Int): String {
+        fun withShoppingListId(shoppingListId: Long): String {
             return route.replace("{shoppingListId}", shoppingListId.toString())
         }
     }
